The int64 iota warned once per fold and was already int32
featurization.py asked broadcasted_iota for int64 and JAX, without x64,
truncated it to int32 and said so -- a six-line UserWarning across every
notebook's output, naming a line the reader cannot act on.

They are sort indices over the last axis, so int32 is not a concession.
Measured rather than assumed: the int64 request comes back with dtype int32
and one warning, the int32 request comes back with the same dtype, no
warning, and the permutations are element-for-element identical.

esm.py has the other jnp.int64 in the tree and is left alone: it declares an
input aval for lowering rather than an op's dtype, it is on the ESM path
only, and it is not what was reported.

+++ src/alphafold3/model/network/featurization.py
@@ -110,7 +110,10 @@ def gumbel_argsort_sample_idx(
   # This construction is equivalent to jnp.argsort, but using a non stable sort,
   # since stable sort's aren't supported by jax2tf
   axis = len(logits.shape) - 1
-  iota = jax.lax.broadcasted_iota(jnp.int64, logits.shape, axis)
+  # int32, not int64: these are sort indices over the last axis, and JAX
+  # without x64 truncates the request anyway -- which it says, once per
+  # fold, as a UserWarning from this line. Same values, no warning.
+  iota = jax.lax.broadcasted_iota(jnp.int32, logits.shape, axis)
   _, perm = jax.lax.sort_key_val(
       logits + z, iota, dimension=-1, is_stable=False
   )
